"timer registers factored out; updates #1173"
